Fixing a bug where new file could be written to wrong location.
If the file paths given to the script are leaf file names, then the
script would incorrectly build the new file path and put the new file at
the root of the drive rather than in the same directory as the original
file.

This occurred because the original file path was parsed to extract the
directory portion and the leaf portion then use the directory as the
base for the renamed file.  However, if you supply only the leaf to
begin with, directory is an empty string so you end up with "/leaf.ext".